requirements.txt
flask
flask-sqlalchemy
flask-cors
bcrypt
werkzeug
pymysql
cryptography
pyjwt==2.3.0
python-dotenv==0.19.0
requests